refactor: move eval_spline_nodes and eval_rational_nodes to least_square
